+++ frontend/webapp/core/fetch_hijack.py
@@ -6,7 +6,7 @@ from webapp.utils import API_URI
 log = logging.getLogger("webapp_logger")
 
 
-def get_hijack_by_key(hijack_key):
+def check_if_hijack_exists(hijack_key):
     try:
         log.debug("send request for total get_hijack_by_key")
         url_ = API_URI + "/view_hijacks?key=eq." + hijack_key
@@ -14,7 +14,7 @@ def get_hijack_by_key(hijack_key):
         raw_json = response.json()
         log.debug("response: {}".format(raw_json))
         if raw_json:
-            return raw_json[0]
+            return True
     except BaseException:
         log.exception("failed to fetch get_hijack_by_key")
-    return None
+    return False