+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 ### Added
 - Support for dormant flags in hijacks
 - Storing hijack update (origin, neighbor) combinations in redis
+- Learn rule action/button for ongoing hijacks in frontend
+- Translate learn rule request in ARTEMIS-compatible dicts in backend
+- Translate learn rule ARTEMIS-compatible dicts into yaml conf in backend
+- Update yaml conf with learned rule
 - Configured/matched prefix field and search in frontend hijack and update tables
 - Initial kubernetes/helm (helm-charts) support
 - Reject old updates from taps and have a "HISTORIC" variable to enable/disable